Attempt to style the button #1
I'm having an issue with changing the style of a button in the "About" block of the main page. Although It was partially successful, I noticed that while I press the button, it changes color to grey and the text color to black. After using the inspect tool, I found that the issue lies in the SCSS file, which is connected to our site via an external link. I tried to adjust the "focus" pseudo-class, but it didn't work.