Site Navigation Update
Added rounded corners to the images inside the UI card, formatting the rest of the section. Generated AI images using Adobe Firefly and inserted them instead of the placeholder image. Added another folder inside the assets to organise the project, therefore had to change the path for images on the main page.