UI cards update/ Website Functionality
Firstly I added the actual content instead of placeholders, and then I changed the order of cards so they correspond to the menu. A lot of text inside the card looked visually not appealing, so I implemented jQuery function $('.read-more').click() and event.preventDefault() to prevent the default link behavior. toggleReadMore that toggles the visibility of the additional content within each card when the "Read more" button is clicked.