Script update
This updated script includes modifications to the toggleReadMore function. Now, when a "Read more" button is clicked, it first hides any other expanded content and changes the text of other buttons to "Read more". Then, it shows the content for the clicked card and changes its button text to "Read less". This behavior allows only one card's content to be open at any given time, closing the other card's content when a new card is opened.