Logo #2
Got rid of png and substitute them with svg files. In order to create a hover state for the logo i had to resort to js script. This JavaScript code listens for the mouseover event on the .logo-container. When the mouse hovers over the container, it changes the src attribute of the .logo-img to the hover image source. On mouseout, it changes it back to the default image source. This method uses JavaScript event listeners to handle the image swap effect on hover without relying solely on CSS.