Map section Update
Firstly I've added icons, which I created with the Icons8 service. After that they've been linked up to the relevant sources as well as read more button. I've faced an issue with map being shorter then the content on the right side in order to fix it im adjusting the padding-bottom property in the .map-responsive class, it helps me to control the aspect ratio of the map. The map's height will adjust to match the height of the content on the right side of the page. Also i've used the same type of animation for the icons as I did before to the UI cards, using "transform: scale(1.05);" for the class hover state. Visually i added rounded corners to the map with the radius formula as i did to the UI cards, as well as i played around with padding and margins of the elements.