Benefits Section upd
Got rid of useless code, changed a few bits of code to rearrange and fit the content. Placed the content in a .col-md-12.col-lg-6 column to take up half of the row.
Placed the image in another .col-md-12.col-lg-6 column to sit alongside the content column on larger screens.
Ensured that the image will be centered within its column on all screen sizes due to the text-center class.